Several functions added
The setup to send a email is now in the send view controller, it has to
be tested.

The map view controller was added, the idea is to go to that view and
take a screenshot to send it in a email

Two lines where added in the part to setup the email, they are
commented because I need to test it, the function of these lines is to
take attach the screenshot image to the email to send it.